[silicon_creator,hmac] workaround double implementation for hmac functions
This file declares an inline implementation for hmac_sha256_init and
hmac_sha256_final. This file is then imported by the mock C+++
implementation which re-implement these functions with a C++ implementation.

Uses __cplusplus exclusion as a temporary workaround, although the real
fix should better be to remove the inline implementation from the .h
file.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>